# internal/sumdb/api/api.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.1 DO NOT EDIT.
package api

import (
	"fmt"
	"net/http"
	"time"

	"github.com/gin-gonic/gin"
	"github.com/oapi-codegen/runtime"
)

// Hash defines model for Hash.
type Hash struct {
	Hash  string `json:"hash"`
	Index int64  `json:"index"`
}

// HashList defines model for HashList.
type HashList = []Hash

// Record defines model for Record.
type Record struct {
	CreatedAt time.Time `json:"createdAt"`
	Data      string    `json:"data"`
	Id        int64     `json:"id"`
	Path      string    `json:"path"`
	UpdatedAt time.Time `json:"updatedAt"`
	Version   string    `json:"version"`
}

// RecordList defines model for RecordList.
type RecordList = []Record

// Tree defines model for Tree.
type Tree struct {
	CreatedAt   time.Time `json:"createdAt"`
	Name        string    `json:"name"`
	Size        int64     `json:"size"`
	UpdatedAt   time.Time `json:"updatedAt"`
	VerifierKey string    `json:"verifierKey"`
}

// TreeList defines model for TreeList.
type TreeList = []Tree

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// List available sumdb trees
	// (GET /api/v1/sumdb/trees)
	ListTrees(c *gin.Context)
	// List hashes in the specified tree
	// (GET /api/v1/sumdb/trees/{name}/hashes)
	ListTreeHashes(c *gin.Context, name string)
	// List records in the specified tree
	// (GET /api/v1/sumdb/trees/{name}/records)
	ListTreeRecords(c *gin.Context, name string)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandler       func(*gin.Context, error, int)
}

type MiddlewareFunc func(c *gin.Context)

// ListTrees operation middleware
func (siw *ServerInterfaceWrapper) ListTrees(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.ListTrees(c)
}

// ListTreeHashes operation middleware
func (siw *ServerInterfaceWrapper) ListTreeHashes(c *gin.Context) {

	var err error

	// ------------- Path parameter "name" -------------
	var name string

	err = runtime.BindStyledParameterWithOptions("simple", "name", c.Param("name"), &name, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter name: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.ListTreeHashes(c, name)
}

// ListTreeRecords operation middleware
func (siw *ServerInterfaceWrapper) ListTreeRecords(c *gin.Context) {

	var err error

	// ------------- Path parameter "name" -------------
	var name string

	err = runtime.BindStyledParameterWithOptions("simple", "name", c.Param("name"), &name, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter name: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.ListTreeRecords(c, name)
}

// GinServerOptions provides options for the Gin server.
type GinServerOptions struct {
	BaseURL      string
	Middlewares  []MiddlewareFunc
	ErrorHandler func(*gin.Context, error, int)
}

// RegisterHandlers creates http.Handler with routing matching OpenAPI spec.
func RegisterHandlers(router gin.IRouter, si ServerInterface) {
	RegisterHandlersWithOptions(router, si, GinServerOptions{})
}

// RegisterHandlersWithOptions creates http.Handler with additional options
func RegisterHandlersWithOptions(router gin.IRouter, si ServerInterface, options GinServerOptions) {
	errorHandler := options.ErrorHandler
	if errorHandler == nil {
		errorHandler = func(c *gin.Context, err error, statusCode int) {
			c.JSON(statusCode, gin.H{"msg": err.Error()})
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandler:       errorHandler,
	}

	router.GET(options.BaseURL+"/api/v1/sumdb/trees", wrapper.ListTrees)
	router.GET(options.BaseURL+"/api/v1/sumdb/trees/:name/hashes", wrapper.ListTreeHashes)
	router.GET(options.BaseURL+"/api/v1/sumdb/trees/:name/records", wrapper.ListTreeRecords)
}